+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 1.0.3
+
+Change switch and checkbox names
+
+## 1.0.2
+
+Change package description
+
+## 1.0.1
+
+Add detailed package description
+
 ## 1.0.0
 
 First release